Make more refactoring and add tests for product service

+++ 3. Software development/2. Connect to DB/Product_Web/Product_Web/Constants/ProductConstants.cs
@@ -2,6 +2,6 @@
 {
     public static class ProductConstants
     {
-        public const decimal VAT = (decimal)0.2;
+        public const decimal VAT = 0.2m;
     }
 }
